User stuff

+++ SysProjectBackend/src/main/java/facade/DataFacade.java
@@ -121,7 +121,7 @@ public class DataFacade {
         }
     }
     
-    public UserDTO editUser(User u, long id) {
+    public UserDTO editUser(User u, int id) {
         EntityManager em = emf.createEntityManager();
         User tmp = null;
         try {
